Default build-system and license can be specified in global config (#256)

+++ rye/src/cli/init.rs
@@ -4,27 +4,17 @@ use std::str::FromStr;
 use std::{env, fs};
 
 use anyhow::{anyhow, bail, Context, Error};
-use clap::{Parser, ValueEnum};
+use clap::Parser;
 use console::style;
 use license::License;
 use minijinja::{context, Environment};
 use pep440_rs::VersionSpecifier;
-use serde::Serialize;
 
 use crate::config::Config;
 use crate::platform::{get_default_author, get_latest_cpython, get_python_version_from_pyenv_pin};
+use crate::pyproject::BuildSystem;
 use crate::sources::PythonVersion;
 use crate::utils::is_inside_git_work_tree;
-
-#[derive(ValueEnum, Copy, Clone, Serialize, Debug)]
-#[value(rename_all = "snake_case")]
-#[serde(rename_all = "snake_case")]
-pub enum BuildSystem {
-    Hatchling,
-    Setuptools,
-    Flit,
-    Pdm,
-}
 
 /// Creates a new python project.
 #[derive(Parser, Debug)]
@@ -44,9 +34,9 @@ pub struct Args {
     /// Do not create .python-version file (requires-python will be used)
     #[arg(long)]
     no_pin: bool,
-    /// Which build system should be used?
-    #[arg(long, default_value = "hatchling")]
-    build_system: BuildSystem,
+    /// Which build system should be used(defaults to hatchling)?
+    #[arg(long)]
+    build_system: Option<BuildSystem>,
     /// Which license should be used (SPDX identifier)?
     #[arg(long)]
     license: Option<String>,
@@ -185,25 +175,26 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
     );
     let version = "0.1.0";
     let author = get_default_author();
-    let license = if let Some(license) = cmd.license {
-        if !license_file.is_file() {
-            let license_obj: &dyn License = license
-                .parse()
-                .expect("current license not an valid license id");
-            let license_text = license_obj.text();
-            let rv = env.render_named_str(
-                "LICENSE.txt",
-                LICENSE_TEMPLATE,
-                context! {
-                    license_text,
-                },
-            )?;
-            fs::write(&license_file, rv)?;
-        };
-        Some(license)
-    } else {
-        None
+    let license = match cmd.license {
+        Some(license) => Some(license),
+        None => cfg.default_license(),
     };
+    if license.is_some() && !license_file.is_file() {
+        let license_obj: &dyn License = license
+            .clone()
+            .unwrap()
+            .parse()
+            .expect("current license not an valid license id");
+        let license_text = license_obj.text();
+        let rv = env.render_named_str(
+            "LICENSE.txt",
+            LICENSE_TEMPLATE,
+            context! {
+                license_text,
+            },
+        )?;
+        fs::write(&license_file, rv)?;
+    }
 
     // write .python-version
     if !cmd.no_pin && !python_version_file.is_file() {
@@ -229,6 +220,11 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
         false
     };
 
+    let build_system = match cmd.build_system {
+        Some(build_system) => build_system,
+        None => cfg.default_build_system().unwrap_or(BuildSystem::Hatchling),
+    };
+
     let rv = env.render_named_str(
         "pyproject.json",
         TOML_TEMPLATE,
@@ -239,7 +235,7 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
             requires_python,
             license,
             with_readme,
-            build_system => cmd.build_system,
+            build_system,
         },
     )?;
     fs::write(&toml, rv).context("failed to write pyproject.toml")?;
